requirements.txt
streamlit
pathlib
nsfw-detector
opencv-python==4.6.0.66
pillow
numpy
tensorflow-cpu

